MyBlog/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-extensions==3.2.3
django-taggit==5.0.1
djangorestframework==3.15.1
httpie==3.2.2
idna==3.6
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
oauthlib==3.2.2
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
PySocks==1.7.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==69.2.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
transliterate==1.10.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
